DS614-Faculty-Recommender/requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
scikit-learn
requests
beautifulsoup4
sentence-transformers
faiss-cpu
google-generativeai